remove RC namespace
