+++ lib/elements/column/base.rb
@@ -1,0 +1,6 @@
+module Elements
+  module Column
+    class Base
+    end
+  end
+end
