--- lib/elements/rc/column/base.rb
@@ -1,16 +1,0 @@
-module Elements
-  module RC
-    module Column
-      class Base
-        attr_reader :length_x, :length_y, :height, :material
-
-        def initialize(length_x:, length_y:, height:, material:)
-          @length_x = length_x.to_f
-          @length_y = length_y.to_f
-          @height = height.to_f
-          @material = material
-        end
-      end
-    end
-  end
-end
