--- lib/elements/rc/column/rectangular.rb
@@ -1,10 +1,0 @@
-require 'elements/rc/column/base'
-
-module Elements
-  module RC
-    module Column
-      class Rectangular < Base
-      end
-    end
-  end
-end
