--- lib/elements/rc/footing/isolated.rb
@@ -1,10 +1,0 @@
-require 'elements/rc/footing/base'
-
-module Elements
-  module RC
-    module Footing
-      class Isolated < Base
-      end
-    end
-  end
-end
