+++ spec/elements/column/rectangular_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+require 'elements/column/rectangular'
+require 'materials/concrete'
+
+RSpec.describe Elements::Column::Rectangular do
+  xit 'does something'
+end
