--- spec/elements/rc/column/rectangular_spec.rb
@@ -1,7 +1,0 @@
-require 'spec_helper'
-require 'elements/rc/column/rectangular'
-require 'materials/concrete'
-
-RSpec.describe Elements::RC::Column::Rectangular do
-  xit 'does something'
-end
