+++ spec/engineering/analysis/footing/centric_isolated_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 require 'engineering/analysis/footing/centric_isolated'
-require 'elements/rc/column/rectangular'
-require 'elements/rc/footing/isolated'
+require 'elements/column/rectangular'
+require 'elements/footing'
 require 'loads/point_load'
 
 RSpec.describe Engineering::Analysis::Footing::CentricIsolated do
@@ -10,10 +10,10 @@ RSpec.describe Engineering::Analysis::Footing::CentricIsolated do
   end
 
   let(:column) do
-    Elements::RC::Column::Rectangular.new(length_x: 500, length_y: 300, height: 2900, material: 'concrete')
+    Elements::Column::Rectangular.new(length_x: 500, length_y: 300, height: 2900, material: 'concrete')
   end
   let(:footing) do
-    Elements::RC::Footing::Isolated.new(length_x: 4000, length_y: 4000, height: 400, material: 'concrete')
+    Elements::Footing.new(length_x: 4000, length_y: 4000, height: 400, material: 'concrete')
   end
   let(:load_from_column) { Loads::PointLoad.new(value: 1500, location: nil) }
   let(:effective_height) { 250 }
